Extracted inner loops into separate functions and added by iterations and by tolerance functions
The by_iterations implementation allows the caller to specify the number of iterations that they desire to perform of a givne algorithm to solve the system.
The by_tolerance functions allows the caller to specify a given tolerance value, and the program will continue iterating until the error is reduced within the specified tolerance.
TODO : Maybe there are edge cases where the by_tolerance functions could iterate for extended periods of time. To solve this, maybe add a version with an upper limit number of allowed iterations to prevent this issue on said edge cases.